+++ storage-provider/server/src/pipeline/mod.rs
@@ -11,7 +11,7 @@ use polka_storage_proofs::{
 };
 use polka_storage_provider_common::rpc::ServerInfo;
 use primitives::{
-    commitment::{CommD, CommP, CommR, Commitment},
+    commitment::{CommP, Commitment},
     proofs::derive_prover_id,
     randomness::{draw_randomness, DomainSeparationTag},
     sector::SectorNumber,
@@ -20,7 +20,7 @@ use storagext::{
     types::{
         market::DealProposal,
         storage_provider::{
-            PartitionState, PoStProof, ProveCommitSector, SectorPreCommitInfo,
+            PartitionState, PoStProof, ProveCommitSector,
             SubmitWindowedPoStParams,
         },
     },
@@ -41,8 +41,6 @@ use types::{
 
 use crate::db::{DBError, DealDB};
 
-// TODO(@th7nder,#622,02/12/2024): query it from the chain.
-const SECTOR_EXPIRATION_MARGIN: u64 = 20;
 
 #[derive(Debug, thiserror::Error)]
 pub enum PipelineError {
@@ -337,121 +335,22 @@ async fn precommit(
 ) -> Result<(), PipelineError> {
     tracing::info!("Starting pre-commit");
 
-    let sealer = Sealer::new(state.server_info.seal_proof);
-    let Some(mut sector) = state.db.get_sector::<UnsealedSector>(sector_number)? else {
+    let Some(sector) = state.db.get_sector::<UnsealedSector>(sector_number)? else {
         tracing::error!("Tried to precommit non-existing sector");
         return Err(PipelineError::SectorNotFound);
     };
-    // Pad sector so CommD can be properly calculated.
-    sector.piece_infos = sealer.pad_sector(&sector.piece_infos, sector.occupied_sector_space)?;
-    tracing::debug!("piece_infos: {:?}", sector.piece_infos);
 
-    tracing::info!("Padded sector, commencing pre-commit and getting last finalized block");
-
-    let current_block = state.xt_client.height(true).await?;
-    tracing::info!("Current block: {current_block}");
-
-    let digest = state
-        .xt_client
-        .get_randomness(current_block)
-        .await?
-        .expect("randomness to be available as we wait for it");
-
-    let entropy = state.xt_keypair.account_id().encode();
-    // Must match pallet's logic or otherwise proof won't be verified:
-    // https://github.com/eigerco/polka-storage/blob/af51a9b121c9b02e0bf6f02f5e835091ab46af76/pallets/storage-provider/src/lib.rs#L1539
-    let ticket = draw_randomness(
-        &digest,
-        DomainSeparationTag::SealRandomness,
-        current_block,
-        &entropy,
-    );
-
-    let cache_path = state.sealing_cache_dir.join(sector_number.to_string());
+    let cache_dir_path = state.sealing_cache_dir.join(sector_number.to_string());
     let sealed_path = state.sealed_sectors_dir.join(sector_number.to_string());
-    tokio::fs::create_dir_all(&cache_path).await?;
-    tokio::fs::File::create_new(&sealed_path).await?;
 
-    // TODO(@th7nder,31/10/2024): what happens if some of the process fails? SP will be slashed, and there is no error reporting? what about retries?
-    let sealing_handle: JoinHandle<Result<PreCommitOutput, _>> = {
-        let prover_id = derive_prover_id(state.xt_keypair.account_id());
-        let cache_dir = cache_path.clone();
-        let unsealed_path = sector.unsealed_path.clone();
-        let sealed_path = sealed_path.clone();
+    let sector = sector.pre_commit(
+        state.xt_client.clone(),
+        &state.xt_keypair,
+        cache_dir_path,
+        sealed_path
+    ).await?;
 
-        let piece_infos = sector.piece_infos.clone();
-        tokio::task::spawn_blocking(move || {
-            sealer.precommit_sector(
-                cache_dir,
-                unsealed_path,
-                sealed_path,
-                prover_id,
-                sector_number,
-                ticket,
-                &piece_infos,
-            )
-        })
-    };
-    let sealing_output = sealing_handle.await??;
-    tracing::info!(
-        "Created sector's replica, CommD: {}, CommR: {}",
-        sealing_output.comm_d.cid(),
-        sealing_output.comm_r.cid()
-    );
-
-    let sealing_output_commr = Commitment::<CommR>::from(sealing_output.comm_r);
-    let sealing_output_commd = Commitment::<CommD>::from(sealing_output.comm_d);
-
-    tracing::debug!("Precommiting at block: {}", current_block);
-    let result = state
-        .xt_client
-        .pre_commit_sectors(
-            &state.xt_keypair,
-            vec![SectorPreCommitInfo {
-                deal_ids: sector.deals.iter().map(|(id, _)| *id).collect(),
-                expiration: sector
-                    .deals
-                    .iter()
-                    .map(|(_, deal)| deal.end_block)
-                    .max()
-                    .expect("always at least 1 deal in a sector")
-                    + SECTOR_EXPIRATION_MARGIN,
-                sector_number: sector_number,
-                seal_proof: state.server_info.seal_proof,
-                sealed_cid: sealing_output_commr.cid(),
-                unsealed_cid: sealing_output_commd.cid(),
-                seal_randomness_height: current_block,
-            }],
-            true,
-        )
-        .await?
-        .expect("we're waiting for the result");
-
-    let precommited_sectors = result
-        .events
-        .find::<storagext::runtime::storage_provider::events::SectorsPreCommitted>()
-        // `.find` returns subxt_core::Error which while it is convertible to subxt::Error as shown
-        // it can't be converted by a single ? on the collect, so the type system tries instead
-        // subxt_core::Error -> PipelineError
-        .map(|result| result.map_err(|err| subxt::Error::from(err)))
-        .collect::<Result<Vec<_>, _>>()?;
-
-    let sector = PreCommittedSector::create(
-        sector,
-        cache_path,
-        sealed_path,
-        sealing_output_commr,
-        sealing_output_commd,
-        current_block,
-        precommited_sectors[0].block,
-    )
-    .await?;
     state.db.save_sector(sector.sector_number, &sector)?;
-
-    tracing::info!(
-        "Successfully pre-commited sectors on-chain: {:?}",
-        precommited_sectors
-    );
 
     state
         .pipeline_sender
